Fix variable type for registerCpu if needed
If registerCpu is set in the config, the previous code would set it in the database as a string preventing the user from running any code. This seems like a good way to block a user :) This change will update the variable to a number if it is not already a number.